requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.20.0
keras==3.12.0
numpy
pillow
